refactor(map): address local review P2s
- extract pure style module layers/map-style.ts; map-layers.tsx is now just
  the MapLayers component
- single ageCase() helper replaces five hand-rolled age-category case
  expressions; collapses to scalar when uniform, skips fallback-equal branches
- basemapWashOpacity joins MapStyleKnobs; map-tiles.ts reads it, panel
  copy-values round-trips a plain MapStyleKnobs
- panel sizes polygon-proxy-points by proxy count (was detailed-point count);
  takes features and computes counts itself, devPointCount removed from
  full-map
- AGE_PALETTES moved into the dev panel, its only consumer
