+++ src/components/dataset/map/style-tuning-panel.tsx
@@ -1,19 +1,20 @@
 "use client";
 
-import { useCallback, useEffect, useState } from "react";
+import { useCallback, useEffect, useMemo, useState } from "react";
 import { useMap } from "react-map-gl/maplibre";
+import type { Feature } from "geojson";
 import {
-  AGE_PALETTES,
   DEFAULT_STYLE_KNOBS,
-  buildAgeColorExpression,
-  buildAgeOpacityExpression,
+  ageCase,
   buildLineWidth,
   buildPointRadiusForCount,
   buildPointStrokeWidth,
   buildPolygonStrokeWidth,
+  type AgeCategoryColors,
   type AgeCategoryValues,
   type MapStyleKnobs,
-} from "./layers/map-layers";
+} from "./layers/map-style";
+import { createSmallPolygonProxyPoints } from "./layers/polygon-proxy-points";
 
 type AgeCategory = keyof AgeCategoryValues<number>;
 
@@ -24,12 +25,100 @@ const AGE_CATEGORIES: { key: AgeCategory; label: string }[] = [
   { key: "very-old", label: "Very old" },
 ];
 
+// Candidate age palettes for the palette dropdown. All keep contrast against
+// the washed basemap; the high-contrast group keeps chroma on every step so
+// the very-old majority never fades out
+const AGE_PALETTES: Record<string, AgeCategoryColors> = {
+  // CARTO BluGrn (previous default; pale light end lost the very-old dots)
+  "blu-grn": {
+    recent: "#1d4f60",
+    medium: "#36877a",
+    older: "#6dbc90",
+    "very-old": "#c4e6c3",
+  },
+  // Hand-built single-hue teal, palest light end
+  "teal-ramp": {
+    recent: "#0f5c54",
+    medium: "#2f9e8f",
+    older: "#7fb8ad",
+    "very-old": "#cbd5d1",
+  },
+  // ColorBrewer BuGn (blue-green, warmer light end)
+  bugn: {
+    recent: "#238b45",
+    medium: "#66c2a4",
+    older: "#b2e2e2",
+    "very-old": "#e5f5f9",
+  },
+  // CARTO DarkMint (deeper, bluer dark end)
+  "dark-mint": {
+    recent: "#123f5a",
+    medium: "#3a7c89",
+    older: "#7bbcb0",
+    "very-old": "#d2fbd4",
+  },
+  // The four below skip the palest ramp steps so the very-old majority
+  // keeps enough chroma to stay visible against the washed basemap
+  // CARTO Emrld (saturated green light end)
+  emrld: {
+    recent: "#105965",
+    medium: "#4c9b82",
+    older: "#6cc08b",
+    "very-old": "#97e196",
+  },
+  // CARTO Teal (steely, most muted light end of this group)
+  teal: {
+    recent: "#2a5674",
+    medium: "#4f90a6",
+    older: "#85c4c9",
+    "very-old": "#a8dbd9",
+  },
+  // ColorBrewer YlGnBu (blue dark end, warm green-yellow light end)
+  ylgnbu: {
+    recent: "#225ea8",
+    medium: "#41b6c4",
+    older: "#7fcdbb",
+    "very-old": "#c7e9b4",
+  },
+  // ColorBrewer GnBu (strong blue dark end, minty light end)
+  gnbu: {
+    recent: "#08589e",
+    medium: "#4eb3d3",
+    older: "#7bccc4",
+    "very-old": "#a8ddb5",
+  },
+  // Viridis picks: multi-hue, colorblind safe, no step washes out
+  // (the baked default)
+  viridis: DEFAULT_STYLE_KNOBS.colors,
+  // CARTO Sunset picks: purple recent -> amber old, warm against cool basemap
+  sunset: {
+    recent: "#5c53a5",
+    medium: "#a059a0",
+    older: "#eb7f86",
+    "very-old": "#fac484",
+  },
+  // Context-vs-accent: old features are visible neutral slates, color is
+  // spent entirely on recency (brand-adjacent greens)
+  "slate-accent": {
+    recent: "#14532d",
+    medium: "#15803d",
+    older: "#64748b",
+    "very-old": "#94a3b8",
+  },
+  // ColorBrewer Blues compressed into the mid-dark range; very-old is a
+  // full mid-tone, not a tint
+  blues: {
+    recent: "#08306b",
+    medium: "#2171b5",
+    older: "#6baed6",
+    "very-old": "#9ecae1",
+  },
+};
+
 // Dev-only live style tuner: writes straight to the map with
 // setPaintProperty, so knob changes restyle instantly without a recompile.
 // Tuned values are exported via "Copy values" and pasted back into
-// DEFAULT_STYLE_KNOBS in map-layers.tsx.
-
-const DEFAULT_WASH_OPACITY = 0.4;
+// DEFAULT_STYLE_KNOBS in map-style.ts.
 
 // Dev-only UI, intentionally not translated
 const TEXT = {
@@ -44,7 +133,7 @@ const TEXT = {
 };
 
 type StyleTuningPanelProps = {
-  pointCount: number;
+  features: Feature[];
 };
 
 function Knob({
@@ -123,13 +212,25 @@ function Group({
   );
 }
 
-function StyleTuningPanelInner({ pointCount }: StyleTuningPanelProps) {
+function StyleTuningPanelInner({ features }: StyleTuningPanelProps) {
   const { current: mapRef } = useMap();
   const [open, setOpen] = useState(false);
   const [knobs, setKnobs] = useState<MapStyleKnobs>(DEFAULT_STYLE_KNOBS);
-  const [washOpacity, setWashOpacity] = useState(DEFAULT_WASH_OPACITY);
   const [zoom, setZoom] = useState<number | null>(null);
   const [copied, setCopied] = useState(false);
+
+  // Density-aware radius needs the same counts the baked layers use: point
+  // features for detailed-points, proxy centroids for polygon-proxy-points
+  const { pointCount, proxyPointCount } = useMemo(() => {
+    const polygonFeatures = features.filter(
+      (f) =>
+        f.geometry.type === "Polygon" || f.geometry.type === "MultiPolygon"
+    );
+    return {
+      pointCount: features.filter((f) => f.geometry.type === "Point").length,
+      proxyPointCount: createSmallPolygonProxyPoints(polygonFeatures).length,
+    };
+  }, [features]);
 
   const update = useCallback((mutate: (draft: MapStyleKnobs) => void) => {
     setKnobs((prev) => {
@@ -160,9 +261,9 @@ function StyleTuningPanelInner({ pointCount }: StyleTuningPanelProps) {
         map.setPaintProperty(layer, prop as never, value as never);
       }
     };
-    const ageColor = buildAgeColorExpression(knobs.colors);
+    const ageColor = ageCase(knobs.colors);
 
-    setPaint("basemap-mute", "background-opacity", washOpacity);
+    setPaint("basemap-mute", "background-opacity", knobs.basemapWashOpacity);
 
     setPaint(
       "detailed-points",
@@ -170,11 +271,7 @@ function StyleTuningPanelInner({ pointCount }: StyleTuningPanelProps) {
       buildPointRadiusForCount(pointCount, knobs)
     );
     setPaint("detailed-points", "circle-color", ageColor);
-    setPaint(
-      "detailed-points",
-      "circle-opacity",
-      buildAgeOpacityExpression(knobs.point.opacity)
-    );
+    setPaint("detailed-points", "circle-opacity", ageCase(knobs.point.opacity));
     setPaint(
       "detailed-points",
       "circle-stroke-width",
@@ -186,7 +283,7 @@ function StyleTuningPanelInner({ pointCount }: StyleTuningPanelProps) {
     setPaint(
       "polygon-proxy-points",
       "circle-radius",
-      buildPointRadiusForCount(pointCount, knobs)
+      buildPointRadiusForCount(proxyPointCount, knobs)
     );
     setPaint(
       "polygon-proxy-points",
@@ -208,10 +305,10 @@ function StyleTuningPanelInner({ pointCount }: StyleTuningPanelProps) {
     setPaint("aoi-boundary", "line-color", knobs.boundary.color);
     setPaint("aoi-boundary", "line-width", knobs.boundary.width);
     setPaint("aoi-boundary", "line-opacity", knobs.boundary.opacity);
-  }, [mapRef, knobs, washOpacity, pointCount]);
+  }, [mapRef, knobs, pointCount, proxyPointCount]);
 
   const copyValues = () => {
-    const payload = JSON.stringify({ washOpacity, ...knobs }, null, 2);
+    const payload = JSON.stringify(knobs, null, 2);
     navigator.clipboard.writeText(payload);
     setCopied(true);
     setTimeout(() => setCopied(false), 1500);
@@ -246,11 +343,11 @@ function StyleTuningPanelInner({ pointCount }: StyleTuningPanelProps) {
       <Group title="Basemap" defaultOpen>
         <Knob
           label="Wash opacity"
-          value={washOpacity}
+          value={knobs.basemapWashOpacity}
           min={0}
           max={1}
           step={0.05}
-          onChange={setWashOpacity}
+          onChange={(v) => update((k) => (k.basemapWashOpacity = v))}
         />
       </Group>
 
@@ -474,10 +571,7 @@ function StyleTuningPanelInner({ pointCount }: StyleTuningPanelProps) {
           {copied ? TEXT.copied : TEXT.copy}
         </button>
         <button
-          onClick={() => {
-            setKnobs(DEFAULT_STYLE_KNOBS);
-            setWashOpacity(DEFAULT_WASH_OPACITY);
-          }}
+          onClick={() => setKnobs(DEFAULT_STYLE_KNOBS)}
           className="flex-1 rounded border border-gray-300 px-2 py-1 text-xs hover:bg-gray-50"
         >
           {TEXT.reset}
